highlight obj
everything is in this highlight prefab. it checks if the distance between the obj and the player is in radius then it will beep.

default setting is that the animation will be in sync with the rest. uncomment the animation line if u want to make it to not sync with the rest of the item.

do not save this prefab in another prefab. only drag and drop in the scene hierarchy